Create stream_ordered_memory_resource base class.
